requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==2.0.12
colorama==0.4.6
Django==4.2.20
django-filter==25.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.10
flake8==7.1.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.1
mccabe==0.7.0
packaging==24.2
pluggy==1.0.0.dev0
py==1.11.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.26.0
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
